fix: rename Reservation button component
